Core/Objects: Create a base class for Object to support creating new client visible object types
