+++ src/server/game/Entities/Object/ObjectGuid.h
@@ -81,7 +81,7 @@ enum TypeMask
     TYPEMASK_WORLDOBJECT            = TYPEMASK_UNIT | TYPEMASK_GAMEOBJECT | TYPEMASK_DYNAMICOBJECT | TYPEMASK_CORPSE | TYPEMASK_AREATRIGGER | TYPEMASK_SCENEOBJECT | TYPEMASK_CONVERSATION
 };
 
-inline constexpr std::array<uint32, NUM_CLIENT_OBJECT_TYPES> ObjectTypeMask =
+inline constexpr std::array<uint32, NUM_CLIENT_OBJECT_TYPES + 1> ObjectTypeMask =
 {
     TYPEMASK_OBJECT,
     TYPEMASK_OBJECT | TYPEMASK_ITEM,
@@ -101,6 +101,7 @@ inline constexpr std::array<uint32, NUM_CLIENT_OBJECT_TYPES> ObjectTypeMask =
     TYPEMASK_OBJECT | TYPEMASK_AI_GROUP,
     TYPEMASK_OBJECT | TYPEMASK_SCENARIO,
     TYPEMASK_OBJECT | TYPEMASK_LOOT_OBJECT,
+    0,
 };
 
 enum class HighGuid
